requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.0.6
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
flake8==7.1.1
idna==3.10
kombu==5.4.2
mccabe==0.7.0
mysqlclient==2.2.4
prompt_toolkit==3.0.48
psycopg2==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
